IPCV/Code/requirements.txt
numpy
pillow
matplotlib
tk
